Close popup windows on outside mouse press and release
Popups are expected to automatically close, when the user clicks outside
of them. Clicking on a different window should close the popup window,
regardless of whether that window is owned by the same application or
not.

In qt widgets, we use QWindow::setMouseGrabEnabled(true) and
QWindow::setKeyboardGrabEnabled(true), to receive and
act upon these mouse events, that would otherwise be sent to different
applications. While this approach doesn't work on every single platform,
it should work Windows, macOS and Linux/xcb, which are the most
important ones. Closing popups in Wayland unfortunately can't be done in
the same manner. The solution here was to close all popups on the
ApplicationDeactivate signal (914c2ef4fb926c16e8874c9c3691401f7e22ae22).

In qt quick, we have the closePolicy property, which we should honor
to some degree. QWindow::setMouseGrabEnabled(true), gives us the
flexibility needed to do that, and we can do the grabbing in
QQuickPopupWindowPrivate::setVisible(), instead of static Q*Application
functions which is the pattern used by widgets.

Since we're already forwarding mouse and keyboard events to the active
popup window, we'll need to close popups in cases when the user presses
or releases the mouse outside of the popup window.
The filterPopupSpecialCases() function already handles some special cases
when the popup window is used by a menu. Instead of reimplementing the
mousePressEvent and mouseReleaseEvent functions in QQuickPopupWindow,
I instead made the decision to extend the functionality of
filterPopupSpecialCases to handle cases for closing all sorts of popups
when the mouse is pressed or released outside, instead of just limiting
this functionality to menus. Note that instead of closing all popups,
when clicking outside, we're only closing the current active popup
window. This is consistent with qt widgets. Strangely however, the popup
order functions in a FIFO manner, where the first opened popup is the
active popup, and not the last. This means that popups are not ordered
in a stack, but rather a queue.

In cases when the closePolicy is CloseOnEscape, we don't have to do
anything, since escape key handling is already taken care of,
by the forwarding logic.

Pick-to: 6.8
Change-Id: Icc1cfcf1526108e1bf7be90bac608fb094aa8f7b
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>