PopupPositioner: ensure popup windows are placed inside a screen
When `popupType: Popup.Item`, the popup positioner will attempt to flip,
move and resize the popup item, in order to make it fit inside its
window.

We want something similar for when `popupType: Popup.Window`, except
that the bounds in this case, ends up being the geometry of the screen
that contains the popup window, instead of the nearest parent item's
window.

The omission of any resizing to the popup window, is done on purpose.
Since the popup positioner's current role, is to only affect the popup
window position. This might change in the future, in case we decide to
refactor some of the logic around popup window resizing.

Pick-to: 6.9 6.8
Change-Id: Iee0f1c7363624b73d0b64f0b6c47765abdb365e2
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>