qmllint: Introduce qds plugin with ErrFunctionsNotSupportedInQmlUi
Introduce the qds warning plugin for .ui.qml file and implement the
illicit warning category that warns about calling methods that can't
be used in .ui.qml files called FunctionsNotSupportedInQmlUi.

Keep the same warning text and name as in qmljscheck from Qt Creator to
not regress and confuse users when changing to qmlls.

Add a special case for Date methods: we can't resolve the Date method so
we have to whitelist all methods that could be called on Date.

In the QtC codemodel, the warning category
ErrFunctionsNotSupportedInQmlUi emits errors when defining QML methods
and when calling functions and in this commit only the function calling
functionality is implemented via warnings instead of errors.

Emitting warnings when defining QML methods is currently not possible
with our current QQmlSA infrastructure and might be added in a future
commit.

Task-number: QTBUG-129308
Change-Id: I4527a90faced4259be72e2dac437b87ee2ba3214
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>