Cleanup pointer grabs when subscene agents are used
When dispatching a pointer event at the end of the method if the event
is a pointer end event we clear up any potential grabbers on the
pointing device.

If a subscene delivery agent exist, or ever existed, we go down a
different path dispatching the initial event there. This path has an
early return which can skip the clearGrabbers line of the main function.

Pick-to: 6.8
Change-Id: Ib441ac09baa5667fe2e6af64d593f7956a3e2f86
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>