Implement QQuickGridViewPrivate::getAttachedObject()
QQuickGridView inherits from QQuickItemView, which provides
functions that emit the signals "pooled" and "reused".
The signals are not emitted if the virtual function
QQuickItemViewPrivate::getAttachedObject() hasn't been implemented
for the appropriate subclass of QQuickItemViewPrivate,
which in this case is QQuickGridViewPrivate.

Implement QQuickGridViewPrivate::getAttachedObject() to enable emitting
signals for when items are being pooled and reused.

Fixes: QTBUG-122249
Fixes: QTBUG-107458
Change-Id: Ibb7351c6fd01ef304ecdcab8344f2b6543ae2777
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>