Sync QQuickWindowQmlImpl::visible with QWindow
QQuickWindowQmlImpl introduces its own `visible' Q_PROPERTY so that it
can do its own things (such as applyWindowVisibility()). However, things
get out of sync if the user of the window uses show() instead of setting
visible=true.

An example of such a case is included as a new unit test for
QQuickWindow.

Pick-to: 6.8 6.9
Change-Id: I0b5b2635919f61a50999fc3ba15a87de8f7074e9
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>