Android: add onStatusChanged() overload passing a QtQuickViewContent
Add new onStatusChanged() method to QtQmlStatusChangeListener
interface which receives the QtQuickViewContent with it.
This offers the users a way to identify which QtQuickViewContent
sent the onStatusChanged event. This is useful if there are
multiple QtQuickViewContents in the app.

Change protected getLibraryName, getModuleName and getFilePath
methods of QtQuickViewContent.java to public, as they are needed
for user to extract information out of the QtQuickViewContent object
received with the onStatusChanged.

Fixes: QTBUG-127420
Change-Id: I1dea9af32670e32fbe5bac7b7bcd16dde2341032
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>