Don't block press and release events for non-popup children of overlay
When trying to position items above an active popup, users will
typically have to reparent those items to the overlay. But doing so, can
break input handling when using modal popups. Since
childMouseEventFilter will cause the overlay to filter events, sent to
it's children, and modal popups assumes it's unsafe to not block those
events if the target item doesn't belong to itself.

It appears the assumption is that only popup items (and dimmers) will be
parented to the overlay, but in reality, any item can, by assigning the
parent to be `Overlay.overlay`. If the item is rendered above a modal
popup, then maybe that popup shouldn't filter the event, similar to the
exception it makes for its own child items.

Pick-to: 6.8 6.5
Task-number: QTBUG-129520
Fixes: QTBUG-130331
Change-Id: Id7b69570eb6598d1ef030e0b7230489046444e45
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>