Qt Quick for Android: Implement new onStatusChanged in examples
Change currently used onStatusChanged method to new one which delivers
the QtQuickViewContent with it in qtquickview  examples.
To showcase how to use the new onStatusChanged method and how it's
useful in an application where there is multiple Qt Quick view contents.

Add private variable for current Qt Quick view content and store
the content coming with onStatusChanged in it. To avoid unnecessary
code execution and warnings in other parts of the example.

Task-number: QTBUG-127420
Change-Id: I375d21611c071d0d1b17cef7f51f221b21c74e09
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>