SourceLocation: deduplicate row/column - offset mapping logic
Move the logic that maps rows and columns to offsets (and backwards) to
SourceLocation static helper methods, and reuse them in qmllsutils.cpp.
This avoids the code duplication that happened in SourceLocation's
endZeroLengthLocation().

Also fix a bug where \r\n was confused with \n\r: \r\n is a windows
newline and counts as 1 instead of 2 newlines.

Fix a test that was confusing '\n\r' and `\r\n` (as the implementation
in qqmllsutils was buggy) and add two tests to avoid miscounting
newlines again.

Change textOffsetFrom() and textRowAndColumnFrom() in QQmlLSUtils to use
the implementation in SourceLocation and adapt the indexes (one starts
at 0 while the other one starts at 1).

Fix some tests in tst_qmlls_utils that were testing weird and invalid
edge cases that treats \r like a normal character and \n like a
character that can't be returned. Treat both \r and \n like characters
whose offset can be returned.

Adapt qmlls range formatting to the new behavior of testOffsetFrom()
that returns \n instead of \r when asking for the a character behind the
end of file. Add tests for files ending without newline and for
files with windows newlines. The files with windows newlines has two
expected outcomes: the formatted parts has \r\n newlines on windows and
\n newlines otherwise.

Pick-to: 6.8
Change-Id: I0ef42c1fc7073eb5fe25ceb893a5d24e976e45fc
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>