DialogsQuickImpl: Implement ColorInputs as a custom control
ColorInputs might as well be a real custom control, instead of a
half-baked one.

This also makes the ColorDialog much more responsive when resizing
it.

If we refactor the module in the future, we'll also likely need each
style to be self-contained to one module each. This patch ensures that
ColorInputs.qml is no longer shared between all styles.

Pick-to: 6.9
Fixes: QTBUG-106246
Task-number: QTBUG-104091
Change-Id: I6c8a687a798f015dce47cc94a10266227899db0f
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>