Move DelegateChooser from Qt.labs.models to QtQml.Models
I assume the intention was to eventually move it out of Qt.labs.models,
once we've had a second look at the API, and are happy enough with it.

The DelegateChooser uses the adaptor model's `value` function, to fetch
data from the model, which it then uses for finding a suitable
DelegateChoice to use. (Based on the DelegateChoice::roleValue
property).

Since the adaptor model's `value` function takes a QString, as the role
to read data from, it won't make sense to change the datatype of
DelegateChooser::role.

The rest of the API makes sense too, in my opinion, and I see no reason
to change it.

Add a simple test in tst_qqmldelegatemodel, since QQmlDelegateModel is
using the DelegateChooser in QQmlDelegateModelPrivate::resolveDelegate.

[ChangeLog][QtQmlModels][Important Behavior Change] DelegateChooser has
been moved from Qt.labs.models to QtQml.Models.

Fixes: QTBUG-98101
Task-number: QTBUG-100069
Change-Id: I76ea58f57475b6a069517b5030e2ba3ac47007d2
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>