qmllint plugins: prefix settingsname with plugin name
Prefix the logging category name with the plugin's name, in the same way
as the category id's are also prefixed.

Fixed a test from bc709184afcb8373be6d69309cb53294455d7248 that did not
test anything. Add a flag to callQmllint() that enables to read settings,
and make sure that a warning is actually emitted when the settings are
not read for the settings/plugin/elementPass_pluginTest.qml file.

The latter file was also renamed because the test lint plugin only
enables itself on files ending with "pluginTest.qml".

Add a compatibility mode that allows loading "old" .qmllint.ini files
where the settings names do not contain the plugin name.

[ChangeLog][Important Behavior Change] Qmllint prefixes logging
categories from plugins with the plugin name in .qmllint.ini files.
For example, PropertyChangesParsed=disable becomes
Quick.PropertyChangesParsed=disable.

Fixes: QTBUG-130357
Change-Id: I47a0bf22991d6a438aa3371c666da34f2c2835bc
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>