sg: Limit buffer pool sizes
Begin limiting the number of buffers pushed to m_vboPool and m_iboPool
when a batch is removed. The important part here is that this mechanism
has no direct equivalent in Qt 5, where existing OpenGL buffers got
respecified (i.e., calling glBufferData with whatever the new size is)
aggressively. Qt 6 has better potential for reusing existing native
buffers as-is (which should be good both for performance and possibly
reducing memory fragmentation). However, this has the potential to
change the applications' memory usage patterns, although this depends
heavily on the scene and how it changes over time.

Begin keeping track of the total size of the buffers pushed to the
vbo/ibo lists. If the a buffer does not fit, just destroy it.

By default the limit is set to 2 MB, quite low compared to the previous
no limit approach. However, for the typical 2D Quick scene with small
geometries (and so vertex/index buffers) this still allows quite a lot
of reuse, while avoiding sky high memory use in scenes that generate
huge geometries and then do not do much with the scene, meaning no
new batches are created in the renderer and so the big buffers just
sit in the pool waiting for reuse in the future... (or until
QQuickWindow::releaseResources() is called)

The limit applies separately for m_vboPool and m_iboPool, and can
be overridden by setting QSG_RENDERER_BUFFER_POOL_LIMIT to the limit
(in bytes; 0 is also valid, disabling buffer reuse completely).

There is a new renderer debug category 'pools' that can be enabled
by setting QSG_RENDERER_DEBUG=pools. This is because in applications
with huge geometries categories like 'upload' are useless, since the
debug output will choke on all the output. So to just see how these
pools grow/shrink, is is better to have a dedicated one as well.

Pick-to: 6.8
Change-Id: I6cc5eb32759c9d10a77f6dcf9b5814613039e208
Reviewed-by: Andy Nichols <andy.nichols@qt.io>