aotstats: Add integration test
The behavior of aotstats has been fragile in the past. Cement it by
adding a test for the expected output for the stats of a project. The
test covers different potentially tricky scenarios.

The output is now sorted alphabetically for module URI and filename.
Bindings and functions are already sorted by file order. This makes the
output deterministic and also more easy to navigate.

Task-number: QTBUG-124667
Pick-to: 6.8
Change-Id: Ie5917c53b3826d39b044e2fb2763a5af2106bccf
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>