QtQml: Make LoadHelper behave like a proper QQmlDataBlob
It should load on the type loader thread so that we can uphold the
invariants on what methods are called on what thread. This way we can
also meaningfully loadFromModule in asynchronous mode. It turns out we
also don't need to store the loaded type separately since the LoadHelper
already holds it.

Fixes: QTBUG-108746
Change-Id: Id0c0bc5f80763c238711f09873153a94cec5244e
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>