CMake: Enable Qml import plugins finalizer during the repo build
The qml import plugins finalizer was only enabled for an installed
qtdeclarative, by appending it to the
INTERFACE_QT_EXECUTABLE_FINALIZERS property of the target when the
ConfigExtras.cmake file is loaded.
That file is not loaded during the build of qtdeclarative, so we
never had finalizers run during the build.

Add the finalizer also during the build, so that when qmlimportscanner
is available, we can also run the finalizer for things like tests.

We currently don't run finalizers for tests on all platforms, but this
change is a pre-requisite for that.

Pick-to: 6.8
Task-number: QTBUG-93625
Task-number: QTBUG-112212
Change-Id: Ib5d2a80766232001712b9e7fab7da640f7947c51
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>