qqmljssourcelocation: fix wrong asserts on 32 bit systems
Replace the asserts comparing 0-based offsets and 1-based columns and lines:
the first character of a file has offset 0 but line and column 1, so the
current assert is always false.

Also, the assert tests an assumption which does not hold:
for example the DOM formatting code works with "local" sourcelocation
that has "local" offsets but global line and column number.

Instead, use a more relaxed check and make sure that line and column are
in-bounds.

Also, the members of SourceLocation are not protected with setters, so
anyone can write anything there anyway.

Amends 4baa3b7c5b5f6a2e89f701651e6240e2a17cc38b that introduced the
assert and expected the CI to complain on them.

Change-Id: I330b9542e1b123bc3df401414d1c9a22504ba01b
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>