CMake: Make Qml look for qmlsc rather than Qt6QmlCompilerPlusPrivate
When qmlsc is installed, the Qml package tries to look for it by
loading the Qt6QmlCompilerPlusPrivate package. But in addition to
creating the qmlsc imported target, that package also creates a static
QmlCompilerPlusPrivate library, which in turns has dependencies on
other static libraries like QmlCompiler and QmlToolingSettings.

In a user project, Qml has no reason to look for those libraries, it's
only interested in the qmlsc tool.

This change modifies the find_package call to look for the
Qt6QmlCompilerPlusPrivateTools package instead, which brings in the
qmlsc tool, and doesn't have any other dependencies on other library
packages.

It also restores a previous attempt of properly handling
cross-compilation, where we need to look in the host path when it is
set. This is similar to the code we have in
qtdeclarative/src/CMakeLists.txt.

This shouldn't cause the previous issue that led to the revert,
because we don't look for the library packages anymore, only the tool
package, so it's safe to manipulate the CMAKE_PREFIX_PATH and friends.

Reverts 090f07a503989f93d714a44b522a8acf0921b845
Adjusts 80cb4091d6cc5b7d13b80668f6e393788d763b3e

Pick-to: 6.8
Fixes: QTBUG-130534
Task-number: QTBUG-105488
Change-Id: Ie6530f642441db0a61687318a7426a82f4c512d3
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>