QQuickWidget: Update initial contentItem size
QQuickwidget manually keeps the window contentItem size in sync with the
window whenever the window resizes. This is needed as resizeEvents are
not sent for non-platform windows which normally takes care of the sync.

The initial setGeometry call changes the window size without the
contentItem size. This leads to the wrong contentItem size until the
widget resizes later.

Fixes: QTBUG-128935
Pick-to: 6.8
Change-Id: Ib184b5d0ced7c3609705abb3ad8e3d621a39a761
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>