Don't propagate mouse event for menu configured inside popup
The patch e4a159632ec39542ec23567963191162f9dcdfef fixes propagating
mouse events by not allowing it to the underlying items beneath the
overlay.

It still can cause an issue when the menu is nested within the popup.
This is because the menu doesn't block the input, which allows the
overlay to pass on the events to the next child item.

This patch enables the menu to consume the mouse press event (as
similar to the drawer), further causing overlay not to pass on the
events to the other child items.

Fixes: QTBUG-130536
Pick-to: 6.8 6.5
Change-Id: I1014c692a4a5aed22e096a79c06e27cb26fea70d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>