CMake: Add a global QT_QML_NO_CACHEGEN cmake variable
A new QT_QML_NO_CACHEGEN cmake variable can be set to disable
compilation of qml files into bytecode or C++ code.

The variable can be set either in a specific directory scope or in
the root of the project (to disable qml compilation for all targets).

This is just a more global version of the NO_CACHEGEN option that can
be passed to qt6_add_qml_module.

For projects with many qml files, this reduces the build time,
allowing faster development iteration times.
It is more convenient than passing NO_CACHEGEN to each created
qml module.

[ChangeLog][Build System] A new QT_QML_NO_CACHEGEN cmake variable can
be set to disable compilation of qml files for all project qml files.

Pick-to: 6.8
Task-number: QTBUG-125631
Task-number: QTBUG-126205
Task-number: QTBUG-128323
Change-Id: I963c14ad0d07bf1f6d6205afdd0a8a90e73f025d
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>