QQuickApplication: do not leak screen info on updateScreens()
There is potential memory leak on updating a list of QQuickScreenInfo
objects in QQuickApplication::updateScreens(). When updating the list
we try to reuse existent screen info objects by simply resizing the
list and creating new objects when needed. But we forgot to delete
unused objects when the list shrinks.

To fix this we just wrap QQuickScreenInfo pointers to std::unique_ptr
in order to enable automatic memory management.

Change-Id: Ib07afe66f269863f728824903c0b50b6206a05e8
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>