QQuickScreenInfo: make sure we handle the wrapped screen changes
QQuickScreenInfo is created every time we access the Window's `screen`
property. See QQuickWindowQmlImpl::screen(). However, it didn't handle
any changes to the screen passed to the constructor. Therefore, any
bindings to the internal properties of the Window's screen didn't work
correctly either.

Fix this by calling setWrappedScreen() directly from the constructor
to ensure that we handle all changes to the wrapped platform's screen.

Fixes: QTBUG-130689
Pick-to: 6.8 6.5
Change-Id: I92337169c12c2c294ee00af45f3885f53b5e6588
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>