Engine: validate that C++ defined singletons have no required properties
Required properties on a singleton don't really make sense. So far we
only detected the issue for composite types, now we only detect it on
(most) C++ defined singletons.
To reduce the overhead of the check, and to avoid weird interactions
with dynamic metaobjecs, we only attempt it when the singleton has a
property cache.

Fixes: QTBUG-122784
Change-Id: Ia4e727d9e2b9171ae6d144acaca3b47cd03f5641
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>