QmlFormat. A bit risky. Remove LineWriterOptions::updateOptions
After QTBUG-128866 it seems that all times we use `LineWriterOptions::Update::None` meaning that not any of the if-s or flags checks on the LineWriterOptions are applicable.
At least for the current production usages (QmlLS, qmlformat)

This opens a great possibility of removing it completely along with other aux entities (see patches in the chain)

Thoughts: Probably in the beginning it was intended to have a capability of changing items / expressions etc. in order not to reparse / rebuild DOM again after reformatting. AFIK atm we do not support any incremental parsing or changes in the DOM / QmlLS, but also this functionality was hard to deal with, therefore it's arguably a better option to remove it.

Change-Id: I00725f012f1860eafd5609e4352bce4880f3ccb3
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>