qqmldomastcreator: fix sourcelocations of bindings
Use the sourcelocation of the binding instead of the property definition
when creating bindings. Add a test.

The behavior prior to this commit leads to a bug in a commit in the
relation chain when linking comments to code, as the comment gets linked
to the Binding instead of the property definition.

Pick-to: 6.8
Change-Id: Id03451967c2093a252c47fe448e5bb46e0358567
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>