QmlDom: simplify AttachedInfo and FileLocations lookup
1. move fileLocations specific lookup from generic AttachedInfo
Initially AttachedInfo intended to be more "generic" structure.
However currently it's specific to the FileLocations field.
Move it to the dedicated function.

2. remove redundant AttachedInfoLookupResultBase
This is a redundant wrapper, since we work only with
the foundTree member, which is essentially an AttachedInfo

3. merge FileLocations::findAttachedInfo into FileLocations::treeOf
In all production cases and in general, we'd like to use foundTree.
Mostly even just particular FileLocations. Hence merging it into
treeOf makes sense

4. remove AttachedInfoLookupResult

there is no reason currently for such an unnecessary complex
wrapper

Task-number: QTBUG-130710
Change-Id: I7ec11ce27512647b3224101aff5190467a1e0337
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>