Add some documentation for ReferenceObject
ReferenceObject is currently the backbone of the the write-back
mechanism.

Add some basic documentation on its purpose and usages to try and reduce
the complexity of approaching some of the dependent code that deals with
write-backs (e.g Sequence, QQmlValueTypeWrapper, ...) and to simplify
approaching a review or performing work related to write-backs, which
currently still require some development.

Change-Id: I56f346c29c79313ebc6181ab87b9f7369115ad83
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>